Some optimizations and benchmarks

+++ src/model/repo_stats.rs
@@ -5,7 +5,7 @@ use std::path::PathBuf;
 pub struct RepoStats {
     pub path: PathBuf,
     pub commits_amount: usize,
-    pub added: usize,
-    pub deleted: usize,
+    pub added: u64,
+    pub deleted: u64,
     pub commits: Vec<Commit>,
 }
